requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
SQLAlchemy==1.4.39
python-dotenv==0.20.0
